feat(array): expose labeled-array axes as read-only properties (a.axis)
Replaces the .size(.axis) method-call sugar for labeled
multi-dimensional
arrays with a direct, parenless a.width / a.height read-only property -
the same no-parens convention already used by arr.length and struct/enum
req getters. .size(.axis) is dropped outright rather than kept
alongside:
every real call site (this repo and whisper-boring) passed a literal
.axis
argument, and the old mechanism resolved it by pattern-matching that
literal
AST node at the call site, not through a real runtime enum value, so
there
was never a parametric use case to preserve.

Rust compiler side:
- desugar_labeled_array.rs: Field-access resolution replaces the removed
  MethodCall("size", [DotIdent]) handling, for both CPU locals and
dynamic
  kernel fields.
- interpreter/eval_gpu.rs: same swap for fixed-shape kernel fields.
- transpiler/{cuda,metal,rocm,wgpu}/device.rs:
try_labeled_array_method_call
  -> try_labeled_array_field_access across all four GPU backends.
- Migrated every real usage to the new syntax: tests_labeled_array.rs,
  desugar_labeled_array/tests.rs, tests/{cuda,rocm,wgpu}_codegen.rs, and
  whisper-boring/src/{audio_gpu,math_gpu}.br.
- docs/array-multidim-types.md, book.md, gpu-module.md updated,
including a
  "History" note on why .size(.axis) was replaced rather than kept.

Self-hosted interpreter (boring/interpreter/*.br):
- First port of the whole labeled multi-dimensional array feature to the
  Boring-written interpreter: ast.br (LabeledAxis, Type.LabeledArray,
  LabeledIndex/LabeledArrayComp/RelabelCast), value.br
(Value.LabeledArray),
  parser_core.br + parser_exprstmt.br
(type/indexing/comprehension/relabel
  parsing), eval.br (evaluation, axis-property read, index assignment),
  exec.br, methods.br (.reshape()/.flatten()).
- Since this interpreter is a dynamically-typed tree-walker with no
checker
  or desugar pass, the design carries the shape on the runtime value
itself
  rather than porting the Rust compiler's static desugar machinery.
- tests/cases/labeled_array.br + .expected, registered in
  tests/interpreter_functional.rs.
- Verified independently of the main crate by transpiling+building the
  interpreter as its own generated Cargo sub-project.

Note: this working tree also carries an in-progress, unrelated Float32/
Float64 width-types migration from a concurrent session touching many of
the same files (boring/interpreter/*.br, src/transpiler/*, etc.) - not
authored in this commit's message above, just along for the ride since
both sessions share this working tree.

+++ src/interpreter/tests.rs
@@ -459,7 +459,7 @@ struct Fahrenheit:
 let c = Celsius(100.0)
 let _result = c as float
 "#;
-    assert_eq!(run_src(src), Value::Float(100.0));
+    assert_eq!(run_src(src), Value::Float64(100.0));
 }
 
 #[test]
@@ -715,7 +715,7 @@ let _result = weight + 1.0
 "#;
     let (interp, res) = run(src);
     res.expect("no runtime error");
-    assert_eq!(get_var(&interp, "_result"), Value::Float(73.5));
+    assert_eq!(get_var(&interp, "_result"), Value::Float64(73.5));
 }
 
 #[test]
@@ -945,7 +945,7 @@ let f = 42 as bool
     res.expect("cast returns nil on failure");
     assert_eq!(get_var(&interp, "a"), Value::Int(42));
     assert_eq!(get_var(&interp, "b"), Value::Nil);
-    assert_eq!(get_var(&interp, "c"), Value::Float(3.14));
+    assert_eq!(get_var(&interp, "c"), Value::Float64(3.14));
     assert_eq!(get_var(&interp, "d"), Value::Nil);
     assert_eq!(get_var(&interp, "e"), Value::Int(1));
     assert_eq!(get_var(&interp, "f"), Value::Nil);
@@ -1541,7 +1541,7 @@ let _a = c.area
     let (interp, res) = run(src);
     res.expect("no runtime error");
     // area = 3.14159 * 2.0 * 2.0 = 12.56636
-    if let Value::Float(f) = get_var(&interp, "_a") {
+    if let Value::Float64(f) = get_var(&interp, "_a") {
         assert!((f - 12.56636).abs() < 0.001, "expected ~12.566, got {}", f);
     } else {
         panic!("expected Float, got {:?}", get_var(&interp, "_a"));
@@ -1884,7 +1884,7 @@ let s = c as string
 "#;
     let (interp, res) = run(src);
     res.expect("no error");
-    assert_eq!(get_var(&interp, "f"), Value::Float(100.0));
+    assert_eq!(get_var(&interp, "f"), Value::Float64(100.0));
     assert_eq!(get_var(&interp, "s"), Value::Str("100°C".into()));
 }
